LU-2181 tests: test failure on conf-sanity, subtest 45
Disable additional subtest 45 for SLES11 SP2.
As in previous failures seen in LU-2181 the umount command in SLES 11
stats the mountpoint before doing the umount syscall, which leads to hangs
when the MDS is unavailable.
This test failure was only seen in full test runs so wasn't seen and fixed
in the previous patch.

Test-Parameters: clientdistro=sles11sp2 testlist=conf-sanity envdefinitions=SLOW=yes
Signed-off-by: Bob Glossman <bob.glossman@intel.com>
Change-Id: I09af0e6e21c1d823077e3142688fddb71b57c8a2