LU-1606 api: compile and build lustreapi test
To avoid regression of the client api conf-sanity includes a new test 72
to verify the lustreapi can be compiled and linked against.  All the
files in the directory $LUSTRE_TESTS_API_DIR with the extension .c are
compiled and linked. The directory is located in the build tree at
./lustre/tests/clientapi.

One program is initially available in the directory:

  #include <lustre/lustreapi.h>
  int main(int argc, char *argv) {
    struct lov_user_md lum;
    return llapi_file_get_stripe(argv[1], &lum);
  }

Signed-off-by: Richard Henwood <richard.henwood@intel.com>
Change-Id: I0ad6a1671bf7033ec2ad5bc7a82d82e468cd31c2