LU-2563 tests: Print correct names in mgsdevname()
Current mgsdevname() prints names of block devices (or files for loop
devices) regardless of MGS back end types.  This is plainly a
violation of what the function is expected to do.  If an MGS is
ZFS-based, mgsdevname() should print the dataset name instead of the
VDev names.  This patch fixes the regression and makes sure that
mgsdevname() and mgsvdevname() work for both standalone and
MDT-combined MGSs.

Signed-off-by: Li Wei <wei.g.li@intel.com>
Change-Id: I5885af48e811fcd98aa00c5d901db30e5837b505