LU-1897 test: replay-single test_70b dbench not found
dbench is exiting before test_70b complets is work.

dbench is being used as a background load in this test.
In one observed case test was taking a little under 90s but
dbench was exiting at 62s.

120s seems like a good next point to try.

This test could stand to be rewritten to make the background
load less error prone so this is just a temp fix to help test
quality.

Signed-off-by: Keith Mannthey <keith.mannthey@intel.com>
Change-Id: Iba243fe3132a5fa677e5a7bcd09e491727ba092a