LU-2467 ptlrpc: Add a module parameter to suppress OBD_PINGs
This patch introduces a new ptlrpc module parameter, "suppress_pings",
to provide an option to reduce excessive OBD_PINGs in large clusters.
By default, the parameter is zero, giving a behavior identical to
current implementation; if it is non-zero on a node, any MDC or OSC on
the node will not send periodic pings and any MDT or OST on the node
will not evict idle clients.  When a server has a zero
"suppress_pings", its clients must also use a zero value to avoid
being evicted for idling.  When a server has a non-zero
"suppress_pings", there must be an external infrastructure that
notifies the targets on the server of client deaths, via the targets'
"evict_client" procfs entries.  In addition, a highly available
standalone MGS is recommended when suppressing pings, so that clients
are notified (by Imperative Recovery) of server recoveries.

The changes do two basically independent things.  One is initializing
import and export states (i.e., imp_suppress_pings and
exp_obd_chain_timed) according to "suppress_pings", since whether to
ping or not is a property of each import-export pair.  MGC pings can
not be suppressed, because maintaining MGS connections is dictated by
the reliance on Imperative Recovery.

The other thing is changing pinger and import recovery routines to
respect the import property set earlier.  (The export side does not
need any change at all.)  Considering that a client in a large cluster
may have quite a number of (e.g., 10,000) imports, we would like to
keep imports off pinger_imports whenever possible, instead of just
iterating all of them and checking imp_suppress_pings.  Imports need
to be added back to pinger_imports occasionally to initiate
recoveries.  Also, pings are still needed to query last committed
transactions if there are uncommitted requests on an import, so that
resources pinned for replays can be released even if applications
become idle.

Test-Parameters: envdefinitions=MODOPTS_PTLRPC="suppress_pings=1",LOAD_MODULES_REMOTE=true
Change-Id: Iabc84d395c978c3f156c52aebfad83621facb4fe
Signed-off-by: Li Wei <wei.g.li@intel.com>