LU-0000 oss: rename OST_* RPC opcodes to OSS_*
For consistency with MDS_* RPC opcodes, and the thought that the RPC
opcodes relate to the remote OSS node more than the target, rename
the OST_* constants to their OSS_* equivalents.  Keep an OST_* opcode
table for API compatibility reasons, but it should no longer be used.

Fix up other RPC opcode declarations to avoid typedefs.

Signed-off-by: Andreas Dilger <andreas.dilger@intel.com>
Change-Id: I83e98e2778276de91856c5b944396f27773ebbe5