LU-2614 build: Modified lbuild script to build SR-IOV enabled OFED
To enable the compilation of the SR-IOV enabled OFED stack provided
by Mellanox, a different compilation command for the OFED kernel has
to be used. The SR-IOV features allows the usage of virtualized functions
within a infiniband HCA, and assign them to different VMs simultanously.
At the moment neither the 'vanilla' OFED (www.openfabrics.org)
nort the customised OFED from Mellanox support this feature.
A new option '--ofed-vm' has been introduced to switch the compilation
on the compilation of the SR-IOV driver.
The name of the file to be used is currently hard coded. A separte
option will be implented to specify the download location in the next
release.

Code changes in download_ofed, build_kernel_ib

Signed-off-by: frank <Frank.Heckes@intel.com>
Change-Id: Iae8eafda2e7c475e519db6d927bf0c70f23ff70e