LU-2264 test: Create memory load for testing
It can be useful to add extra memory pressure for testing.
Client nodes are likely to use extra resources.

rewrite memhog.c to include an automatic mode.  ./memhog -a will
consume all but 160-128mb of Ram and Swap.  The mode consumes
then exits at this time.  Original behavior is maintained with
the exception of moving the chunk size to 32mb (down from 128m)

This submission is to get this code out for basic review. The
question of how to intergarete this so it is used is still
outstanding.

memhog is quite verbose.  I am wondering about changing
it be be more quite by default and add a -v ops flag.

Signed-off-by: Keith Mannthey <keith.mannthey@intel.com>
Change-Id: I2e52a841ed4f54985c7ec2536d49da5f44c85c57