LU-1199 build: Clean out the build directory
The "build" directory has, over time, become a dumping ground
for any random script or file that a person doesn't know where to put.
And the users of "build" seem to be allergic to subdirectories, so it
has become a mess.

Because of the way that Intel's build farm and git hooks expect to
find certain files in fixed locations I have had to leave a few
symlinks in place that point to the new locations.

Change-Id: I04dc529d4f4060b892e1e4eaa8613bbc3337c414
Signed-off-by: Christopher J. Morrone <morrone2@llnl.gov>
Test-Parameters: fortestonly