LU-1897 test: replay-single test_70b dbench not found
dbench is exiting before test_70b complets is work.

dbench is being used as a background load in this test.
In one observed case test was taking a little under 90s but
dbench was exiting at 62s. Duration was set at 90.

Present dbench starts for 60 seconds then the system sleeps
for 12, then we start a new 60 timer in which we expect
dbench to be running.

The mail issue is the use of Duration for use as both the
timer of the real test and the timer of the background load.

I raised the real duration and started the timer for the real
test before starting dbench such that the real test should
finish before dbench does or about the same time.

I also move the the check for dbench out of the main test loop.
We check the return code of dbench at the end of the test
and we know what duration it was started with.  I don't feel
we need to be super parioid about the background workload
when we run that workload on a timer.

Signed-off-by: Keith Mannthey <keith.mannthey@intel.com>
Change-Id: Iba243fe3132a5fa677e5a7bcd09e491727ba092a