LU-793 ptlrpc: Reconnections should not be refused...
Since not letting clients with pending RPCs to reconnect caused a lot
of problems, this patch allows such a reconnection and would only
refuse to process requests with the same xids that are already being
processed. This patch is a port of LU-793 to b_neo_stable (2.2)
Requests from Andreas were added also.

Xyratex-bug-id: MRP-459
Reviewed-by: Alexey Lyashkov <alexey_lyashkov@xyratex.com>
Reviewed-by: Andriy Skulysh <andriy_skulysh@xyratex.com>
Signed-off-by: Iurii Golovach <iurii_golovach@xyratex.com>
Change-Id: I9be309df0abd3f8b4c2078f7a5b46fad8865e5e7