LU-1199 ldiskfs: Export option macros for external use
The "CONFIG_LDISKFS*" macros are ldiskfs's version of
the in-kernel ext4 CONFIG_EXT4* macros.  Currently any external
package using ldiskfs's header files (i.e. lustre) needs to know
exactly how ldiskfs sets those macros and do the same itself.
When ldiskfs and lustre shared a build system, this was somewhat
sloppy, but also relatively safe.  Now that the build systems
are independant, there is room for easy mistakes.

This patch creates a ldiskfs_definitions.h.in file which
creates ldiskfs_definitions.h at ldiskfs configure time.  It
also patches several of the ext4 header files to include this
file.

The intent is that any external users of ldiskfs's header
files will no longer need to guess at the right options to
set.  Instead, they will be explicitly set.

For now, all of the macros are statically defined in
ldiskfs_definitions.h.  But the intent is that we can also
in the future conditionally set macros as a result of autoconf
tests.

Finally, when a later patch adds a ldiskfs-devel rpm file it
will include the <ldiskfs/ldiskfs_definitions.h>.  That will
give lustre an easy file to test with autoconf and/or include
directly if it needs to adjust its build based on supported
ldiskfs options.  Right now lustre only uses
<ldiskfs/ldiskfs_definitions.h> indirectly as a result of
other ldiskfs headers including it.

Change-Id: I5c35f26ff7cb7dd59a6ab431c965980186acc60a
Signed-off-by: Christopher J. Morrone <morrone2@llnl.gov>