LU-2642 mgs: fix missing error code return in mgs_fs_setu
The GOTO macro calls in mgs_fs_setup doesn't assign rc to the value
passed to it. As a result, if dt_locate_at fails, it will return early
but also return instead of PTR_ERR(root) as intended.

This was caught by PTR_ERR being __must_check as of v3.6.34.

Signed-off-by: Jeff Mahoney <jeffm@suse.com>
Change-Id: Ie4897440d481980de6c4d345d2ab2120b44b32c7