LU-2182 utils: new llapi_layout API implementation
Add a new set of llapi routines for interacting with file layouts that
hide the details of the wire-protocol data structures from the user.
Define an opaque data type lustre_layout_t to abstract the layout of a
lustre file and generic accessor functions to read and write it.  The
following visible interfaces are added to liblustreapi.a:

  llapi_layout_alloc - allocate a new layout
  llapi_layout_file_create - create new file with given layout
  llapi_layout_free - free memory allocated for a layout
  llapi_layout_lookup_bypath - get a file layout given a path
  llapi_layout_lookup_fd - get a file layout given a file descriptor
  llapi_layout_obj_id - get object id associated with a stripe
  llapi_layout_ost_index - get OST index associated with a stripe
  llapi_layout_ost_index_set - set OST index assoicated with a stripe
  llapi_layout_pattern - get RAID pattern of a layout
  llapi_layout_pattern_set - set RAID pattern of a layout
  llapi_layout_pool_name - get pool name of a layout
  llapi_layout_pool_name_set - set pool name of a layout
  llapi_layout_stripe_count - get stripe count of a layout
  llapi_layout_stripe_count_set - set stripe count of a layout
  llapi_layout_stripe_size - get stripe size of a layout
  llapi_layout_stripe_size_set - set stripe size of a layout

The layouts are read and written using fgetxattr() and sgetxattr()
instead of ioctl() to make it easier for architectures like Blue Gene
to function-ship the system calls.

Signed-off-by: Ned Bass <bass6@llnl.gov>
Change-Id: I35fb51055b6438ef3090f43c28a4083a66eaa907