LU-2634 ldiskfs: clear extent flag with fast symlink
Fast symlinks on MDT filesystems formatted with the "extents" feature
appear to be created with the EXT4_EXTENTS_FL. e2fsck considers this
a corruption and deletes the symlink.

This patch will fix the problem in the following two situations:
1. If MDT is still formatted with "-O extents", EXT4_INODE_EXTENTS
flag will be prevented setting to the new fast symlink inodes;
2. If extents feature is disabled but the old data already have
extents flag, EXT4_INODE_EXTENTS flag will be cleared for the old
fast symlink inodes and the change will be written back to disk.

Signed-off-by: Liu Ying <emoly.liu@intel.com>
Change-Id: Ib7c1eee7becfaf2b80805d3f6c13658686ad5699