LU-2190 mountconf: revert previous commit 8e249d43f19
Revert "LU-2190 mountconf: zfs processes mount parameters"

This reverts commit 8e249d43f19a4d227c76e6def41d7eb19e6a207f.

The reverted patch is incompatible with existing ZFS based file systems,
and is arguably the wrong approach to take in general. If a single ZFS
data set parameter is desired, a patch needs to be written such that it
is backwards compatible with existing file systems. This will give sites
a "grace period" in which they can convert their file systems to use the
new configuration format.

With that said, I'd argue that using a single ZFS dataset parameter for
the Lustre configuration parameters makes it much more difficult to
administer, and is a regression on its own; regardless of the lack of
backwards compatibility.

Signed-off-by: Prakash Surya <surya1@llnl.gov>
Change-Id: I3a6f005144106966aec0ef18105ce9639cf1f034