LU-2669 scripts: Fix module list in lnet script
The lnet start up script's module list was incorrect causing the
following command to fail:

   # /etc/init.d/lnet stop

This patch adds missing modules, and rearranges the order certain
modules, to match the current dependencies of Lustre modules.

In particular, the osc module was moved to match it's reliance on
lov, and the following modules were added to the list: ofd, mdt,
osd_zfs, osd_ldiskfs, osp, lod, mdd, fid, and fld.

Signed-off-by: Prakash Surya <surya1@llnl.gov>
Change-Id: I081c5feb89ba83960bdc3ed885193d4c7b7cef93