LU-1770 idl: reserve new OBD_CONNECT feature flags
Reserve protocol feature flags on b2_1 to match the definitions
on the master branch:
- OBD_CONNECT_FLOCK_OWNER
- OBD_CONNECT_LVB_TYPE
- OBD_CONNECT_NANOSEC_TIME
- OBD_CONNECT_LIGHTWEIGHT
- OBD_CONNECT_SHORTIO

The OBD_CONNECT_NANOSEC_TIME was originally added in commit
596816513b0603cb159353e1444c97d8bc3bcacf (on master it was
528feb1793e744c63c2cd137d11418c513eb92d), but this was changed
in master commit f100d508311d16a09040a057438ee06fd9e6d7fe due
to a conflict with OBD_CONNECT_FLOCK_OWNER in a customer patch.

As of the time of this patch, the nanosecond timestamp flag
is not yet in use on any branch, so changing the flag value
does not cause any protocol incompatibility.

Signed-off-by: Andreas Dilger <andreas.dilger@intel.com>
Change-Id: Iae34aeec54cdfeb39fe932aa132a3cac982d1ee5