LU-2158 fsfilt: Rename and modify fsfilt_ext3.c
This change replaces the fsfilt_ext3.c file with it's fsfilt_ldiskfs.c
counterpart. Previously this was done dynamically using a complex
solution of autoconf variables, make targets, and scripting. Now that
fsfilt_ldiskfs.c is a "real" source file in the tree, much of the lvfs
build system complexities are unnecessary and have been removed.

Additionally, this is a precursor to removing the lvfs directory
altogether. When that happens, the fsfilt_ldiskfs.c file can be
moved into the ldiskfs OSD layer as a stop gap solution until its
functionality is no longer needed.

Signed-off-by: Prakash Surya <surya1@llnl.gov>
Change-Id: I76931670ad1f2ce2ada5a305ffde3f40df5dcd7f