LU-2473 ldiskfs: Reorganize ldiskfs kernel patches
This commit makes no changes to the patches, only moving
their location.

For some time we have supported only a single kernel per major
OS release, e.g. RHEL 6.3, or RHEL 6.2, but not both.  While
it is understandable to only support a single kernel, it is
quite another to intentionally destroy the patchset for the previous
kernel every time we upgrade.

This makes the logistics of upgrading an OS release (even a minor
one like RHEL 6.2 to 6.3, or 6.3 to 6.4) quite complicated.

We should really start a new patch series when we update support
for a new kernel.  That way we can leave the patch sets for
previous kernels in place.  We do not need to "support" them, but
at least leaving them unmolested until the users of the previous
patch set have time to upgrade would be greatly appreciated.

This commit attempts to organize the patches into subdirectories
according to the kernel/os in which they were first created.  A
later commit will then add support for RHEL6.4.

Also, fix build/confirmpatches.sh to allow checking
ldiskfs/kernel_patches (a bug prevent checking anything but the
default).

Extend both confirmpatches.sh and clearpatches.sh to work with
series files that point to patches organized into subdirectories.

Change-Id: I6552fc271fff1f00657ba1430c4a1215dea5b530
Signed-off-by: Christopher J. Morrone <morrone2@llnl.gov>