LU-2467 ptlrpc: Add a module parameter to suppress OBD_PINGs
This patch introduces a new ptlrpc module parameter, "suppress_pings",
to provide an option for reducing excessive OBD_PINGs in large
clusters.  By default, the parameter is off (zero), giving a behavior
identical to current implementation; if it is on (non-zero) on a node,
any MDC or OSC on that node will not send keep-alive pings and any MDT
or OST on that node will not automatically evict idle clients.  When a
server has "suppress_pings" turned off, its clients must also turn the
parameter off to avoid being evicted for idling.  When a server has
"suppress_pings" turned on, there must be an external mechanism to
notify the targets on the server of client deaths, via the targets'
"evict_client" procfs entries.  In addition, a highly available
standalone MGS is recommended when suppressing pings, so that clients
are notified (by Imperative Recovery) of server recoveries.

The changes do two basically independent things.  One is initializing
import and export states (i.e., imp_suppress_pings and
exp_obd_chain_timed) according to "suppress_pings", since whether to
ping or not is a property of each import-export pair.  MGC pings can
not be suppressed, because maintaining MGS connections is dictated by
the reliance on Imperative Recovery.  This part could be changed in
the future, for example, to be determined by servers and enforced via
connection flags, without impacting the following part too much.

The other thing is changing pinger and import routines to respect the
import property set earlier.  (The export side does not need any
change at all.)  Pings are still needed to query last committed
transactions if there are uncommitted requests on an import, so that
resources pinned for replays can be released even if applications
become idle.  An early version of this patch removes imports that do
not need to be pinged from pinger_imports and add them back when last
committed transactions are needed or recoveries must be initiated.
This version does not do that because a) the overheads of iterating
through 10,000 imports are not prohibitively large---around 10 ms, b)
adding imports back to pinger_imports requires the global pinger_mutex
to be held, and c) the imp_lock contention added on each import is
small.

Change-Id: Iabc84d395c978c3f156c52aebfad83621facb4fe
Signed-off-by: Li Wei <wei.g.li@intel.com>