LU-2622 obdclass: Use per CPU cl_env cache lists
Instead of serializing access to the cl_env cache structure using a
single global lock, this patch attempts to create per CPU caches for the
cl_env structures. This allows a thread to obtain a cl_env from it's CPU
cache without taking a lock, but has the downside of disabling
preemption while it does this.

TODO: I'm not entirely sure if this patch is safe in its current form. I
worry that a CPU in the critical section is interrupted for another
thread which manipulates the per CPU list. This could potentially cause
unforeseen consequences.

Test-Parameters: fortestonly
Signed-off-by: Prakash Surya <surya1@llnl.gov>
Change-Id: Id49ba6b24bf757c146ac913755e7509ebb8d1ecc