LU-1431 llite: limit readahead growth to inode blocksize
With PTLRPC_MAX_BRW_SIZE being as large as 32MB, it doesn't make sense
for the readahead to increase the window size so dramatically at one
time.  Instead, limit readahead growth by the current inode i_blkbits
value, which currently defaults to 4MB.

If we want to tune the readahead growth so that it matches the actual
cl_max_pages_per_rpc value, then this can be done much more easily by
changing the per-inode i_blkbits value, since this is tied to specific
OSTs with specific RPC settings (which may be different for e.g. local
and remote OSTs), and only needs to be checked once.

Change the ras_*() functions to take the inode as an argument, and
always put it first in the argument list for consistency.

Signed-off-by: Andreas Dilger <andreas.dilger@intel.com>
Change-Id: Iec3ca6c3790999a5b7812af56ea289febf3ebbe5