LU-2240 osd: change FID of /ROOT on zfs
pre-production code used FID_SEQ_LOCAL_FILE for /ROOT. with
ldiskfs that FID turns into IGIF which is mapped to MDT0
permanently. with ZFS original local sequence was used which
makes existing setup incompatibile with DNE. The intention of
the patch is to fix this on existing ZFS setups and replace
FID with one from special FID_SEQ_ROOT sequence which is
mapped to MDT0 as well. For simplicity this is done in few
steps:
 - osd-zfs replaces direntry for /ROOT with the new FID and
   fixes OI so that the new FID is mapped to the same dnode
 - MDD finds all the objects listed in /ROOT and updates
   LinkEA properly
 - MDD removes ./.. which might be stored on a disk on pre
   production setups
 - finally MDD reset LMA on /ROOT with the new FID, later
   this is used to recognize already converted filesystems
   (or ones created with the correct FID from the beginning)
   and skip this conversion code

Signed-off-by: Alex Zhuravlev <alexey.zhuravlev@intel.com>
Change-Id: I03062c6909146f9a3aed72f41c0708f9ef92bb82