LU-2654 ldlm: Safely iterate over list of locks in ldlm_resource_dump
It's possible that while we are printing lock details in
ldlm_resource_dump, the lock is removed from the list,
and so as we continue iterating, we are stuck in a tight loop for
the same lock.
Combat this by using cfs_list_for_each_entry_safe() to save a copy
of previous lock we are supposed to jump to.

Also add cfs_list_for_each_entry_safe to libcfs.

Change-Id: I9dfbe84bd63de3093c99e7379daeca30d02fc875
Signed-off-by: Oleg Drokin <green@whamcloud.com>