LU-2640 dt: thandle flag to indicate transaction rollback
Introduce thandle::th_rollback to trace the transaction rollback.
When the thread wants to rollback some sub-operations, it should
set such flag, then in the OSD layer, it can exactly know whether
the operation is for normal modification or for rollback former
modification.

For example: dio_insert() for rename, it is used for inserting
new target name entry, it also may be for adding the old name
entry back when rename failed and need rollback. For the later
case, the thandle::th_rollback should be set before rollback.

Signed-off-by: Fan Yong <fan.yong@intel.com>
Change-Id: I474d4599ccedf9d225c1bd6cb0ab4d2d015a2818