LU-2763 build: git commit fails to allow long Test-Paramters
While trying to a longer Test-Paramater I found I was
unable to. The paramaters were a total of 120 chars or so.
Regardless of any escaping I tried commits were failed
telling me my lines were over 70 chars long.

It was suggest that we use the commit-msg hook from Master.

With local testing it seems to work just fine.

Signed-off-by: Keith Mannthey <keith.mannthey@intel.com>
Change-Id: I5e7c6dfe6f34340ba53e1c2439d2b454a786bc19