LU-1032 build: Remove include of lustre-version.ac
This is an experimental patch. It renames the lustre-version.ac file to
lustre-version.m4. Also, it creates an m4 macro which is used to setup
the Lustre version variables which is called from configure.ac.

Chances are, cleanup like this will be needed to properly support DKMS
builds. This patch is probably incomplete and/or incorrect in and of
itself, but is meant to demonstrate the type of cleanup the build system
might needs to move forward.

Test-Parameters: fortestonly
Signed-off-by: Prakash Surya <surya1@llnl.gov>
Change-Id: I4890a5718755362d9c2e122ddf02e47777896c64