LU-2775 fid: allow FID-on-OST in fid_seq_is_mdt()
The old 2.1 fid_seq_is_mdt() was only checking for FID_SEQ_OST_MDT0
through FID_SEQ_OST_MDT9 values, but this immediately breaks with
FID-on-OST when enabled on 2.4 MDTs/OSTs.  That will be fixed in
a separate 2.4 patch, but the client should not LASSERT() on a
value that it got from the network.  This will also allow the
client to at least access files on MDT0, even if it cannot handle
files on remote MDTs properly.

Signed-off-by: Andreas Dilger <andreas.dilger@intel.com>
Change-Id: I74bc9198799045b8bd91510cb45e8f8765300c1e